+++ NEWS.md
@@ -1,4 +1,9 @@
-# mmrm 0.2.2.9000
+# mmrm 0.2.2.9001
+
+### New Features
+- Add Kenward-Roger support for spatial covariance structures.
+
+# mmrm 0.2.2
 
 ### New Features